update product create DiscountCode

+++ Infra.Data/Repository/CategoryRepository.cs
@@ -13,6 +13,15 @@ namespace Infra.Data.Repository
             _ctx = ctx;   
         }
 
+        public async Task<bool> CheckExistCategoryByName(string categoryName)
+        {
+            return await _ctx.Categories.AnyAsync(c => c.Name == categoryName);
+        }
+        public async Task<bool> CheckExistCategoryById(int categoryId)
+        {
+            return await _ctx.Categories.AnyAsync(c => c.CategoryId == categoryId);
+        }
+
         public async Task CreateCategory(Category category)
         {
             await _ctx.Categories.AddAsync(category);
@@ -22,13 +31,17 @@ namespace Infra.Data.Repository
         public async Task DeleteCategory(int CategoryId)
         {
             var category = await GetCategoryById(CategoryId);
+            if (category == null) 
+            {
+                throw new Exception("این دسته بندی وجود ندارد");
+            }
             _ctx.Categories.Remove(category);
             await Save();
         }
 
         public IEnumerable<Category> GetAllCategories()
         {
-            return new List<Category>().AsEnumerable();
+            return _ctx.Categories.AsEnumerable();
         }
 
         public async Task<Category?> GetCategoryById(int CategoryId)
@@ -38,9 +51,11 @@ namespace Infra.Data.Repository
             return Category;
         }
 
-        public async Task<bool> GetCategoryByName(string categcoryName)
+        public async Task<Category> GetCategoryByName(string categcoryName)
         {
-            return await _ctx.Categories.AnyAsync(c => c.Name == categcoryName);
+            var category = await _ctx.Categories.FirstOrDefaultAsync(c => c.Name == categcoryName);
+
+            return category;
         }
 
         public async Task Save()
@@ -51,7 +66,9 @@ namespace Infra.Data.Repository
         public async Task UpdateCategory(Category category)
         {
             _ctx.Categories.Update(category);
+
             await Save();
         }
+
     }
 }
